study-mule-curl-deploy-v2: api-policy  の作成と変更を疎通確認した

+++ study-mule-deploy/study-mule-curl-deploy-v2/script/deploy-api-policy.sh
@@ -15,9 +15,9 @@ funcMain() {
   funcCheckInputFile
   funcSortConfigFileList
   funcGetApiInstanceList 'before-deploy'
+  funcGetApiInstanceIdList 'before-deploy'
   funcGetApiPolicyList 'before-deploy'
   funcCheckAnypointPlatform
-  funcGetApiInstanceIdList 'before-deploy'
   funcGetApiPolicyIdList 'before-deploy'
   local deployStatus=0
   local tmpFile
@@ -42,7 +42,7 @@ funcMain() {
     echo "Deployed API policy ID: $(cat "$tmpFile")" >&2
   done
   funcGetApiInstanceList 'after-deploy'
-  funcGetApiPolicyIdList 'after-deploy'
+  funcGetApiPolicyList 'after-deploy'
   funcDiffBeforeAfter
   exit "$deployStatus"
 }
@@ -76,6 +76,7 @@ funcCheckInputFile() {
 
 # 入力ファイルをAPIPolicyの作成順序に並び替える。
 # APIPolicyの作成順序は、API InstanceのassetIdの昇順、API PolicyのassetIdの昇順、API Policyのorderの昇順とする。
+# TODO: order=1が存在する状態で、order=1のAPI Policyを作成または、order=1に変更した場合に、既存のAPI Policyのorderを自動的に後ろにズレる想定だが、実機で確認する
 funcSortConfigFileList() {
   local sortedConfigFileList=()
   local configFile
@@ -94,12 +95,13 @@ funcSortConfigFileList() {
       policyFileName="${configFile##*/}"
       policyAssetId="${policyFileName#api-policy_}"
       policyAssetId="${policyAssetId%_*}"
-      policyOrder="$(jq --row-output '.order // empty' "$configFile")"
+      policyOrder="$(jq --raw-output '.order // empty' "$configFile")"
       printf '%s\t%s\t%s\t%s\n' \
         "$apiAssetId" "$policyAssetId" "$policyOrder" "$configFile"
     done | LC_ALL=C sort -t $'\t' -k1,1 -k2,2 -k3,3n -k4,4
   )
   gConfigFileList=("${sortedConfigFileList[@]}")
+  echo "Info: Sorted config file list: ${gConfigFileList[*]}" >&2
 }
 
 funcCheckAnypointPlatform() {
@@ -122,16 +124,21 @@ funcGetApiInstanceList() {
 
 funcGetApiPolicyList() {
   local dirName="$1"
-  local apiInstanceAssetIdAndVersion="$2"
-  local apiInstanceId="$3"
-  local outputFile="$gLogDir/$dirName/api-policy-list_${apiInstanceAssetIdAndVersion}.json"
-  # 呼び出すAPI: APIポリシーの一覧を取得するAPI
-  # API仕様のURL: https://anypoint.mulesoft.com/exchange/portals/anypoint-platform/f1e97bc6-315a-4490-82a7-23abe036327a.anypoint-platform/api-manager-api/minor/1.0/console/method/%236419/
-  local url="${gAnypointBaseUrl}/apimanager/api/v1/organizations/${gOrgId}/environments/${gEnvId}/apis/${apiInstanceId}/policies"
-  funcCurl -- \
-    --header "Authorization: Bearer $gAnypointAccessToken" \
-    "$url"
-  cp -p  "$gLogDir/${gLogSeq}_curl-res-body.json" "$outputFile"
+  local apiInstanceId
+  for apiInstanceId in "${gApiInstanceIdList[@]}"; do
+    local outputFile="$gLogDir/$dirName/api-policy-list_${apiInstanceId}.json"
+    if [ -f "$outputFile" ]; then
+      echo "Info: Skipping API policy list retrieval for API instance ID: $apiInstanceId (file already exists)" >&2
+      continue
+    fi
+    # 呼び出すAPI: APIポリシーの一覧を取得するAPI
+    # API仕様のURL: https://anypoint.mulesoft.com/exchange/portals/anypoint-platform/f1e97bc6-315a-4490-82a7-23abe036327a.anypoint-platform/api-manager-api/minor/1.0/console/method/%236419/
+    local url="${gAnypointBaseUrl}/apimanager/api/v1/organizations/${gOrgId}/environments/${gEnvId}/apis/${apiInstanceId}/policies"
+    funcCurl -- \
+      --header "Authorization: Bearer $gAnypointAccessToken" \
+      "$url"
+    cp -p  "$gLogDir/${gLogSeq}_curl-res-body.json" "$outputFile"
+  done
 }
 
 funcGetApiInstanceIdList() {
@@ -150,7 +157,9 @@ funcGetApiInstanceIdList() {
       "${gLogDir}/${dirName}/api-instance-list.json")
 
     if [ -z "$apiInstanceId" ]; then
-      gApiInstanceIdList+=('<new>')
+      # gApiInstanceIdList+=('<new>') # TODO: depoy-api-instance.sh と処理を共通化を検討する。
+      echo "error: API instance not found for assetId=$apiAssetId, version=$apiVersion" >&2
+      exit 1
     elif [[ "$apiInstanceId" =~ ^[0-9]{1,10}$ ]]; then
       gApiInstanceIdList+=("$apiInstanceId")
     else
@@ -176,7 +185,7 @@ funcGetApiPolicyIdList() {
     apiPolicyId=$(jq --raw-output \
       --arg groupId "$groupIdOfApiPolicy" \
       --arg assetId "$assetIdOfApiPolicy" \
-      '.[] | select(.groupId == $groupId and .assetId == $assetId) | .id // empty' \
+      '.policies[] | select(.template.groupId == $groupId and .template.assetId == $assetId) | .policyId // empty' \
       "${gLogDir}/${dirName}/api-policy-list_${apiInstanceId}.json")
 
     if [ -z "$apiPolicyId" ]; then
@@ -234,7 +243,13 @@ funcDiffBeforeAfter() {
   diff --recursive \
     --new-file \
     --unified \
-    "${gLogDir}/before-deploy" "${gLogDir}/after-deploy"
+    "${gLogDir}/before-deploy" "${gLogDir}/after-deploy" && true
+  rc=$?
+  # diffコマンドは、差分がある場合は1を返すので、1の場合はエラーではない。
+  if [ $rc -gt 1 ]; then
+    echo "Error: diff command failed with exit code $rc" >&2
+    exit $rc
+  fi
 }
 
 # このファイルが、単体試験にて source コマンドで読み込みまれた場合は、funcMain() を実行しないようにする
